+++ delivery-service/src/main/java/com/winner/client/deliveryservice/delivery/application/reader/HubRouteReader.java
@@ -1,8 +1,8 @@
-package com.winner.client.deliveryservice.delivery.application.port;
+package com.winner.client.deliveryservice.delivery.application.reader;
 
 import com.winner.client.deliveryservice.delivery.application.dto.external.HubRouteInfo;
 import java.util.UUID;
 
-public interface HubRoutePort {
+public interface HubRouteReader {
   HubRouteInfo getHubRoutes(UUID fromHubId, UUID toHubId);
 }
