+++ delivery-service/src/main/java/com/winner/client/deliveryservice/delivery/infrastructure/client/HubClientAdapter.java
@@ -1,7 +1,7 @@
 package com.winner.client.deliveryservice.delivery.infrastructure.client;
 
 import com.winner.client.deliveryservice.delivery.application.dto.external.HubRouteInfo;
-import com.winner.client.deliveryservice.delivery.application.port.HubRoutePort;
+import com.winner.client.deliveryservice.delivery.application.reader.HubRouteReader;
 import com.winner.client.deliveryservice.delivery.infrastructure.client.dto.HubRouteResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class HubClientAdapter implements HubRoutePort {
+public class HubClientAdapter implements HubRouteReader {
 
   private final HubClient hubClient;
